+++ checker_test.go
@@ -9,6 +9,8 @@ import (
 
 	"net"
 
+	"time"
+
 	"github.com/sagikazarmark/healthz"
 )
 
@@ -130,6 +132,36 @@ func TestHTTPChecker_Check_Fail(t *testing.T) {
 	assertCheckerFailed(t, checker)
 }
 
+func TestHTTPChecker_Check_Timeout(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		time.Sleep(5 * time.Nanosecond)
+		w.WriteHeader(http.StatusOK)
+		w.Write([]byte("ok"))
+	}))
+	defer ts.Close()
+
+	checker := healthz.NewHTTPChecker(ts.URL, healthz.WithHTTPTimeout(15*time.Millisecond))
+
+	assertCheckerSuccessful(t, checker)
+}
+
+func TestHTTPChecker_Check_Timeout_Fail(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		time.Sleep(10 * time.Nanosecond)
+		w.WriteHeader(http.StatusOK)
+		w.Write([]byte("ok"))
+	}))
+	defer ts.Close()
+
+	checker := healthz.NewHTTPChecker(ts.URL, healthz.WithHTTPTimeout(3*time.Nanosecond))
+
+	err := checker.Check()
+
+	if err == nil {
+		t.Fatal("Expected error, none received")
+	}
+}
+
 func TestTCPChecker_Check(t *testing.T) {
 	addr := "127.0.0.1:54321"
 
@@ -150,6 +182,38 @@ func TestTCPChecker_Check_Fail(t *testing.T) {
 	checker := healthz.NewTCPChecker(addr)
 
 	err := checker.Check()
+
+	if err == nil {
+		t.Fatal("Expected error, none received")
+	}
+}
+
+func TestTCPChecker_Check_Timeout(t *testing.T) {
+	addr := "127.0.0.1:54321"
+
+	lis, err := net.Listen("tcp", addr)
+	if err != nil {
+		t.Fatalf("Received unexpected error: %+v", err)
+	}
+	defer lis.Close()
+
+	checker := healthz.NewTCPChecker(addr, healthz.WithTCPTimeout(15*time.Millisecond))
+
+	assertCheckerSuccessful(t, checker)
+}
+
+func TestTCPChecker_Check_Timeout_Fail(t *testing.T) {
+	addr := "127.0.0.1:54321"
+
+	lis, err := net.Listen("tcp", addr)
+	if err != nil {
+		t.Fatalf("Received unexpected error: %+v", err)
+	}
+	lis.Close()
+
+	checker := healthz.NewTCPChecker(addr, healthz.WithTCPTimeout(3*time.Nanosecond))
+
+	err = checker.Check()
 
 	if err == nil {
 		t.Fatal("Expected error, none received")