+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- TCP Checker
+- HTTP Timeout option
+- TCP timeout option
+
+
 ## 0.5.0 - 2017-03-09
 
 ### Changed